Added a copy process to move all the device xml files to the dist directory on package
I also added any xml files added to the resources/data directory to  the gitignore file.  Because, it would seem, the copyright changed in 2011 on the device file, and if the xml file is added to the repository it'll get a DMCA takedown notice.  This should keep that headache from happening.